fixed CR notes
