+++ pkg/collector/corechecks/gpu/nvidia/clocks.go
@@ -8,7 +8,6 @@
 package nvidia
 
 import (
-	"errors"
 	"fmt"
 
 	"github.com/NVIDIA/go-nvml/pkg/nvml"
@@ -29,13 +28,9 @@ func newClocksCollector(device ddnvml.SafeDevice) (Collector, error) {
 	// Check first if the device supports clock throttle reasons
 	_, err := device.GetCurrentClocksThrottleReasons()
 
-	if err != nil {
-		var nvmlErr *ddnvml.NvmlAPIError
-		if errors.As(err, &nvmlErr) && (nvmlErr.NvmlErrorCode == nvml.ERROR_NOT_SUPPORTED ||
-			nvmlErr.NvmlErrorCode == nvml.ERROR_FUNCTION_NOT_FOUND) {
-			// Only return unsupported device if the API is not supported or symbol not found
-			return nil, fmt.Errorf("%w: %w", errUnsupportedDevice, err)
-		}
+	if err != nil && ddnvml.IsUnsupported(err) {
+		// Only return unsupported device if the API is not supported or symbol not found
+		return nil, fmt.Errorf("%w: %w", errUnsupportedDevice, err)
 	}
 
 	return &clocksCollector{
