+++ pkg/gpu/safenvml/lib.go
@@ -22,39 +22,6 @@ import (
 	"github.com/DataDog/datadog-agent/pkg/util/flavor"
 )
 
-// NvmlAPIError represents an error when interacting with the NVML API.
-// It wraps nvml.Return values to provide idiomatic error handling in Go.
-type NvmlAPIError struct {
-	// APIName is the name of the API that failed
-	APIName string
-	// NvmlErrorCode is the NVML error code
-	NvmlErrorCode nvml.Return
-}
-
-// Error implements the error interface
-func (e *NvmlAPIError) Error() string {
-	switch e.NvmlErrorCode {
-	case nvml.ERROR_FUNCTION_NOT_FOUND:
-		return fmt.Sprintf("%s symbol not found in NVML library", e.APIName)
-	case nvml.ERROR_NOT_SUPPORTED:
-		return fmt.Sprintf("%s is not supported by the GPU or driver", e.APIName)
-	default:
-		return fmt.Sprintf("NVML API error for %s: %s", e.APIName, nvml.ErrorString(e.NvmlErrorCode))
-	}
-}
-
-// NewNvmlAPIErrorOrNil creates a new NvmlAPIError with the given API name and error code,
-// or returns nil if the error code is nvml.SUCCESS
-func NewNvmlAPIErrorOrNil(apiName string, errorCode nvml.Return) error {
-	if errorCode == nvml.SUCCESS {
-		return nil
-	}
-	return &NvmlAPIError{
-		APIName:       apiName,
-		NvmlErrorCode: errorCode,
-	}
-}
-
 // symbolLookup is an internal interface for checking symbol availability
 type symbolLookup interface {
 	lookup(string) error
